Compiles warning free on Vista64 now. ticket #108
git-svn-id: file:///Users/joaander/devel/hoomd-git-convert/repos/trunk@1206 fa922fa7-2fde-0310-acd8-f43f465a7996

+++ src/data_structures/BondData.h
@@ -104,7 +104,7 @@ class BondData
 		//! Get the number of bonds
 		/*! \return Number of bonds present
 		*/
-		const unsigned int getNumBonds() const { return m_bonds.size(); }
+		const unsigned int getNumBonds() const { return (unsigned int)m_bonds.size(); }
 		
 		//! Get a given bond
 		/*! \param i Bond to access
